refactor(InputValidation): moved to internal (#3371)
