+++ src/deprecated/ChoiceFieldset/ChoiceFieldset.tsx
@@ -3,7 +3,7 @@ import {Box, useSSRSafeId} from '../..'
 import createSlots from '../utils/create-slots'
 import {FormValidationStatus} from '../../utils/types/FormValidationStatus'
 import ValidationAnimationContainer from '../../internal/components/ValidationAnimationContainer'
-import InputValidation from '../../_InputValidation'
+import InputValidation from '../../internal/components/InputValidation'
 import ChoiceFieldsetListItem from './ChoiceFieldsetListItem'
 import ChoiceFieldsetDescription from './ChoiceFieldsetDescription'
 import ChoiceFieldsetLegend from './ChoiceFieldsetLegend'
