+++ src/deprecated/InputField/InputField.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {Autocomplete, Box, Select, TextInput, TextInputWithTokens, useSSRSafeId} from '../../'
-import InputValidation from '../../_InputValidation'
+import InputValidation from '../../internal/components/InputValidation'
 import {ComponentProps} from '../../utils/types'
 import {FormValidationStatus} from '../../utils/types/FormValidationStatus'
 import InputFieldCaption from './_InputFieldCaption'
