+++ src/internal/components/InputValidation.tsx
@@ -1,9 +1,9 @@
 import {AlertFillIcon, CheckCircleFillIcon, IconProps} from '@primer/octicons-react'
 import React from 'react'
-import Box from './Box'
-import Text from './Text'
-import {SxProp} from './sx'
-import {FormValidationStatus} from './utils/types/FormValidationStatus'
+import Box from '../../Box'
+import Text from '../../Text'
+import {SxProp} from '../../sx'
+import {FormValidationStatus} from '../../utils/types/FormValidationStatus'
 
 type Props = {
   id: string
